refactor: move types into file
