--- src/Views/Components/AdminUI/ActionMenu/types.ts
@@ -1,8 +1,0 @@
-/**
- *
- * @unreleased
- */
-export type ActionMenuProps = {
-    menuConfig: Array<{title: string; action: () => void}>;
-    toggle: () => void;
-};
