+++ src/Views/Components/AdminUI/Button/index.tsx
@@ -3,12 +3,22 @@ import React from 'react';
 import cx from 'classnames';
 
 import styles from './style.module.scss';
-import {ButtonProps} from '@givewp/components/AdminUI/Button/types';
 
 /**
  *
  * @unreleased
  */
+
+export type ButtonProps = {
+    variant: 'primary' | 'secondary' | 'danger';
+    size: 'small' | 'large';
+    type: 'button' | 'reset' | 'submit';
+    children: React.ReactNode;
+
+    onClick?: React.MouseEventHandler<HTMLButtonElement>;
+    disabled?: boolean;
+    classname?: 'string';
+};
 
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
     ({children, type = 'button', classname, variant = 'primary', size = 'normal', disabled, ...props}, ref) => (
