--- src/Views/Components/AdminUI/Button/types.ts
@@ -1,15 +1,0 @@
-/**
- *
- * @unreleased
- */
-
-export type ButtonProps = {
-    variant: 'primary' | 'secondary' | 'danger';
-    size: 'small' | 'large';
-    type: 'button' | 'reset' | 'submit';
-    children: React.ReactNode;
-
-    onClick?: React.MouseEventHandler<HTMLButtonElement>;
-    disabled?: boolean;
-    classname?: 'string';
-}
