+++ src/Views/Components/AdminUI/FormElements/index.tsx
@@ -1,19 +1,22 @@
 import React from 'react';
 
 import styles from './style.module.scss';
-import {FormElementProps} from '@givewp/components/AdminUI/FormElements/types';
 
 /**
  *
  * @unreleased
  */
 
+export type FormElementProps = {
+    children: React.ReactNode;
+    onSubmit: React.FormEventHandler<HTMLFormElement>;
+    id: string;
+};
+
 const Form: React.FC<HTMLFormElement | FormElementProps> = ({children, id, onSubmit}) => (
     <form className={styles.form} id={id} onSubmit={onSubmit}>
         {children}
     </form>
 );
-
-//Todo: TextFieldInput SelectFieldOptions Label Dropdown
 
 export {Form};
