--- src/Views/Components/AdminUI/FormElements/types.ts
@@ -1,5 +1,0 @@
-export type FormElementProps = {
-    children: React.ReactNode;
-    onSubmit: React.FormEventHandler<HTMLFormElement>;
-    id: string;
-};
