+++ src/Views/Components/AdminUI/FormPage/index.tsx
@@ -5,12 +5,29 @@ import {joiResolver} from '@hookform/resolvers/joi';
 
 import FormNavigation from '@givewp/components/AdminUI/FormNavigation';
 import {Form} from '@givewp/components/AdminUI/FormElements';
-import {FormPageProps} from '@givewp/components/AdminUI/FormPage/types';
 
 /**
  *
  * @unreleased
  */
+
+export interface FormPageProps {
+    formId;
+    handleSubmitRequest: (formValues) => void;
+    defaultValues;
+    validationSchema;
+    children;
+    pageDetails: {
+        id: number;
+        description: string;
+        title: string;
+    };
+    navigationalOptions: Array<{
+        id: number;
+        title: string;
+    }>;
+    actionConfig: Array<{title: string; action: any}>;
+}
 
 export default function FormPage({
     formId,
