--- src/Views/Components/AdminUI/FormPage/types.ts
@@ -1,21 +1,0 @@
-/**
- *
- * @unreleased
- */
-export interface FormPageProps {
-    formId;
-    handleSubmitRequest: (formValues) => void;
-    defaultValues;
-    validationSchema;
-    children;
-    pageDetails: {
-        id: number;
-        description: string;
-        title: string;
-    };
-    navigationalOptions: Array<{
-        id: number;
-        title: string;
-    }>;
-    actionConfig: Array<{title: string; action: any}>;
-}
